backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
futu-api==9.6.5608
httpx==0.26.0
sqlalchemy==2.0.25
pydantic==2.5.3
python-dotenv==1.0.0
